Updated battle calculations to decimal values. Added extension methods to map AcrobaticsLevel and Rank to decimal multipliers for calculation. Added Battle integration tests. Moved common auth code in integration tests to fixture

+++ src/FlyingShadow.Core/Services/Battle/BattleProcessor.cs
@@ -47,7 +47,7 @@ public class BattleProcessor : IBattleProcessor
         };
     }
 
-    private static string GetStatWinner(double statOne, double statTwo, string shadowOneName, string shadowTwoName)
+    private static string GetStatWinner(decimal statOne, decimal statTwo, string shadowOneName, string shadowTwoName)
     {
         return statOne > statTwo 
             ? shadowOneName
